backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.5
pypdf==4.0.0
docx2txt==0.8
sentence-transformers==2.2.2
pinecone-client==2.2.4
chromadb==0.4.22
weaviate-client==3.25.3
pytest==7.4.3
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.3
pandas==2.1.4
pytest-asyncio==0.21.1
httpx==0.26.0
